WIP RunOpenapitoolsCodegenMain
OpenAPI Generator at commit e56a708531e375bec66c901ae41908f3192a2a9b (Pull request 75)

+++ openapi-generator-utils/treepots/treepots-openapigen-java-rest-assured/src/main/java/org/openapitools/client/model/LongEnum.java
@@ -67,7 +67,7 @@ public enum LongEnum {
 
     @Override
     public LongEnum read(final JsonReader jsonReader) throws IOException {
-      Long value = jsonReader.nextInt();
+      Long value = jsonReader.nextLong();
       return LongEnum.fromValue(String.valueOf(value));
     }
   }
